Fix HTG warnings

+++ Common/DataModel/vtkHyperTreeGridNonOrientedMooreSuperCursor.cxx
@@ -384,7 +384,7 @@ void vtkHyperTreeGridNonOrientedMooreSuperCursor::Initialize(vtkHyperTreeGrid* g
        {
          // Cell has a neighbor to the left
          unsigned int r =
-           grid->GetShiftedLevelZeroIndex(treeIndex, -1, 0, 0);
+           grid->GetShiftedLevelZeroIndex(treeIndex, (unsigned int)-1, 0, 0);
          this->Entries[0].Initialize(grid, r);
        } else if (isOld) {
          this->Entries[0].Reset();
@@ -411,13 +411,13 @@ void vtkHyperTreeGridNonOrientedMooreSuperCursor::Initialize(vtkHyperTreeGrid* g
        {
          // Cell has a neighbor to the south
          unsigned int r =
-           grid->GetShiftedLevelZeroIndex(treeIndex, 0, -1, 0);
+           grid->GetShiftedLevelZeroIndex(treeIndex, 0, (unsigned int)-1, 0);
          this->Entries[1].Initialize(grid, r);
          if (toW)
          {
            // Cell has a neighbor to the southwest
            r =
-             grid->GetShiftedLevelZeroIndex(treeIndex, -1, -1, 0);
+             grid->GetShiftedLevelZeroIndex(treeIndex, (unsigned int)-1, (unsigned int)-1, 0);
            this->Entries[0].Initialize(grid, r);
          } else { // if (toW)
            this->Entries[0].Reset();
@@ -426,7 +426,7 @@ void vtkHyperTreeGridNonOrientedMooreSuperCursor::Initialize(vtkHyperTreeGrid* g
          {
            // Cell has a neighbor to the southeast
            r =
-             grid->GetShiftedLevelZeroIndex(treeIndex, 1, -1, 0);
+             grid->GetShiftedLevelZeroIndex(treeIndex, 1, (unsigned int)-1, 0);
            this->Entries[2].Initialize(grid, r);
          } else { // if (toE)
            this->Entries[2].Reset();
@@ -441,7 +441,7 @@ void vtkHyperTreeGridNonOrientedMooreSuperCursor::Initialize(vtkHyperTreeGrid* g
        {
          // Cell has a neighbor to the west
          unsigned int r =
-           grid->GetShiftedLevelZeroIndex(treeIndex, -1, 0, 0);
+           grid->GetShiftedLevelZeroIndex(treeIndex, (unsigned int)-1, 0, 0);
          this->Entries[3].Initialize(grid, r);
        } else { // if (toW)
          this->Entries[3].Reset();
@@ -465,7 +465,7 @@ void vtkHyperTreeGridNonOrientedMooreSuperCursor::Initialize(vtkHyperTreeGrid* g
          {
            // Cell has a neighbor to the northwest
            r =
-             grid->GetShiftedLevelZeroIndex(treeIndex, -1, 1, 0);
+             grid->GetShiftedLevelZeroIndex(treeIndex, (unsigned int)-1, 1, 0);
            this->Entries[5].Initialize(grid, r); // au lieu de 6
          } else { // if (toW)
            this->Entries[5].Reset();
